+++ tests/testthat/test-connection.R
@@ -1,7 +1,7 @@
 library(testthat)
 
 test_that("Open and close connection", {
-  # Postgresql
+  # Postgresql --------------------------------------------------
   details <- createConnectionDetails(dbms = "postgresql",
                                      user = Sys.getenv("CDM5_POSTGRESQL_USER"),
                                      password = URLdecode(Sys.getenv("CDM5_POSTGRESQL_PASSWORD")),
@@ -10,7 +10,7 @@ test_that("Open and close connection", {
   expect_true(inherits(connection, "DatabaseConnectorConnection"))
   expect_true(disconnect(connection))
   
-  # SQL Server
+  # SQL Server --------------------------------------------------
   details <- createConnectionDetails(dbms = "sql server",
                                      user = Sys.getenv("CDM5_SQL_SERVER_USER"),
                                      password = URLdecode(Sys.getenv("CDM5_SQL_SERVER_PASSWORD")),
@@ -19,7 +19,7 @@ test_that("Open and close connection", {
   expect_true(inherits(connection, "DatabaseConnectorConnection"))
   expect_true(disconnect(connection))
   
-  # Oracle
+  # Oracle --------------------------------------------------
   details <- createConnectionDetails(dbms = "oracle",
                                      user = Sys.getenv("CDM5_ORACLE_USER"),
                                      password = URLdecode(Sys.getenv("CDM5_ORACLE_PASSWORD")),
@@ -28,15 +28,18 @@ test_that("Open and close connection", {
   expect_true(inherits(connection, "DatabaseConnectorConnection"))
   expect_true(disconnect(connection))
   
-  # # RedShift details <- createConnectionDetails(dbms = 'redshift', user =
-  # Sys.getenv('CDM5_REDSHIFT_USER'), password = URLdecode(Sys.getenv('CDM5_REDSHIFT_PASSWORD')),
-  # server = Sys.getenv('CDM5_REDSHIFT_SERVER'), schema = Sys.getenv('CDM5_REDSHIFT_CDM_SCHEMA'))
-  # connection <- connect(details) expect_true(inherits(connection, 'DatabaseConnectorConnection'))
-  # expect_true(disconnect(connection))
+  # RedShift  --------------------------------------------------
+  details <- createConnectionDetails(dbms = "redshift", 
+                                     user = Sys.getenv("CDM5_REDSHIFT_USER"), 
+                                     password = URLdecode(Sys.getenv("CDM5_REDSHIFT_PASSWORD")),
+                                     server = Sys.getenv("CDM5_REDSHIFT_SERVER"))
+  connection <- connect(details)
+  expect_true(inherits(connection, "DatabaseConnectorConnection"))
+  expect_true(disconnect(connection))
 })
 
 test_that("Open and close connection using connection strings with embedded user and pw", {
-  # Postgresql
+  # Postgresql --------------------------------------------------
   parts <- unlist(strsplit(Sys.getenv("CDM5_POSTGRESQL_SERVER"), "/"))
   host <- parts[1]
   database <- parts[2]
@@ -56,7 +59,7 @@ test_that("Open and close connection using connection strings with embedded user
   expect_true(inherits(connection, "DatabaseConnectorConnection"))
   expect_true(disconnect(connection))
   
-  # SQL Server
+  # SQL Server --------------------------------------------------
   connectionString <- paste0("jdbc:sqlserver://",
                              Sys.getenv("CDM5_SQL_SERVER_SERVER"),
                              ";user=",
@@ -69,7 +72,7 @@ test_that("Open and close connection using connection strings with embedded user
   expect_true(inherits(connection, "DatabaseConnectorConnection"))
   expect_true(disconnect(connection))
   
-  # Oracle
+  # Oracle --------------------------------------------------
   port <- "1521"
   parts <- unlist(strsplit(Sys.getenv("CDM5_ORACLE_SERVER"), "/"))
   host <- parts[1]
@@ -91,17 +94,30 @@ test_that("Open and close connection using connection strings with embedded user
   expect_true(inherits(connection, "DatabaseConnectorConnection"))
   expect_true(disconnect(connection))
   
-  # RedShift parts <- unlist(strsplit(Sys.getenv('CDM5_REDSHIFT_SERVER'), '/')) host <- parts[1]
-  # database <- parts[2] port <- '5439' connectionString <- paste0('jdbc:redshift://', host, ':', port,
-  # '/', database, '?user=', Sys.getenv('CDM5_REDSHIFT_USER'), '&password=',
-  # URLdecode(Sys.getenv('CDM5_REDSHIFT_PASSWORD'))) details <- createConnectionDetails(dbms =
-  # 'redshift', connectionString = connectionString) connection <- connect(details)
-  # expect_true(inherits(connection, 'DatabaseConnectorConnection'))
-  # expect_true(disconnect(connection))
+  # RedShift --------------------------------------------------
+  parts <- unlist(strsplit(Sys.getenv("CDM5_REDSHIFT_SERVER"), "/")) 
+  host <- parts[1]
+  database <- parts[2] 
+  port <- "5439" 
+  connectionString <- paste0("jdbc:redshift://", 
+                             host, 
+                             ":", 
+                             port,
+                             "/", 
+                             database, 
+                             "?user=", 
+                             Sys.getenv("CDM5_REDSHIFT_USER"), 
+                             "&password=",
+                             URLdecode(Sys.getenv("CDM5_REDSHIFT_PASSWORD"))) 
+  details <- createConnectionDetails(dbms = "redshift", 
+                                     connectionString = connectionString) 
+  connection <- connect(details)
+  expect_true(inherits(connection, "DatabaseConnectorConnection"))
+  expect_true(disconnect(connection))
 })
 
 test_that("Open and close connection using connection strings with separate user and pw", {
-  # Postgresql
+  # Postgresql --------------------------------------------------
   parts <- unlist(strsplit(Sys.getenv("CDM5_POSTGRESQL_SERVER"), "/"))
   host <- parts[1]
   database <- parts[2]
@@ -115,7 +131,7 @@ test_that("Open and close connection using connection strings with separate user
   expect_true(inherits(connection, "DatabaseConnectorConnection"))
   expect_true(disconnect(connection))
   
-  # SQL Server
+  # SQL Server --------------------------------------------------
   connectionString <- paste0("jdbc:sqlserver://", Sys.getenv("CDM5_SQL_SERVER_SERVER"))
   details <- createConnectionDetails(dbms = "sql server",
                                      connectionString = connectionString,
@@ -125,7 +141,7 @@ test_that("Open and close connection using connection strings with separate user
   expect_true(inherits(connection, "DatabaseConnectorConnection"))
   expect_true(disconnect(connection))
   
-  # Oracle
+  # Oracle --------------------------------------------------
   port <- "1521"
   parts <- unlist(strsplit(Sys.getenv("CDM5_ORACLE_SERVER"), "/"))
   host <- parts[1]
@@ -139,11 +155,22 @@ test_that("Open and close connection using connection strings with separate user
   expect_true(inherits(connection, "DatabaseConnectorConnection"))
   expect_true(disconnect(connection))
   
-  # RedShift parts <- unlist(strsplit(Sys.getenv('CDM5_REDSHIFT_SERVER'), '/')) host <- parts[1]
-  # database <- parts[2] port <- '5439' connectionString <- paste0('jdbc:redshift://', host, ':', port,
-  # '/', database) details <- createConnectionDetails(dbms = 'redshift', connectionString =
-  # connectionString, user = Sys.getenv('CDM5_REDSHIFT_USER'), password =
-  # URLdecode(Sys.getenv('CDM5_REDSHIFT_PASSWORD'))) connection <- connect(details)
-  # expect_true(inherits(connection, 'DatabaseConnectorConnection'))
-  # expect_true(disconnect(connection))
+  # RedShift --------------------------------------------------
+  parts <- unlist(strsplit(Sys.getenv("CDM5_REDSHIFT_SERVER"), "/")) 
+  host <- parts[1]
+  database <- parts[2] 
+  port <- "5439" 
+  connectionString <- paste0("jdbc:redshift://", 
+                             host, 
+                             ":",
+                             port,
+                             "/", 
+                             database) 
+  details <- createConnectionDetails(dbms = "redshift", 
+                                     connectionString = connectionString, 
+                                     user = Sys.getenv("CDM5_REDSHIFT_USER"), 
+                                     password = URLdecode(Sys.getenv("CDM5_REDSHIFT_PASSWORD"))) 
+  connection <- connect(details)
+  expect_true(inherits(connection, "DatabaseConnectorConnection"))
+  expect_true(disconnect(connection))
 })