+++ tests/testthat/test-getTableNames.R
@@ -1,7 +1,7 @@
 library(testthat)
 
 test_that("Get table names", {
-  # Postgresql
+  # Postgresql --------------------------------------------------
   details <- createConnectionDetails(dbms = "postgresql",
                                      user = Sys.getenv("CDM5_POSTGRESQL_USER"),
                                      password = URLdecode(Sys.getenv("CDM5_POSTGRESQL_PASSWORD")),
@@ -11,7 +11,7 @@ test_that("Get table names", {
   expect_true("PERSON" %in% tables)
   disconnect(connection)
 
-  # SQL Server
+  # SQL Server --------------------------------------------------
   details <- createConnectionDetails(dbms = "sql server",
                                      user = Sys.getenv("CDM5_SQL_SERVER_USER"),
                                      password = URLdecode(Sys.getenv("CDM5_SQL_SERVER_PASSWORD")),
@@ -21,7 +21,7 @@ test_that("Get table names", {
   expect_true("PERSON" %in% tables)
   disconnect(connection)
 
-  # Oracle
+  # Oracle --------------------------------------------------
   details <- createConnectionDetails(dbms = "oracle",
                                      user = Sys.getenv("CDM5_ORACLE_USER"),
                                      password = URLdecode(Sys.getenv("CDM5_ORACLE_PASSWORD")),
@@ -31,7 +31,7 @@ test_that("Get table names", {
   expect_true("PERSON" %in% tables)
   disconnect(connection)
 
-  # Sqlite
+  # Sqlite --------------------------------------------------
   dbFile <- tempfile()
   details <- createConnectionDetails(dbms = "sqlite",
                                      server = dbFile)
@@ -42,12 +42,13 @@ test_that("Get table names", {
   disconnect(connection)
   unlink(dbFile)
   
-  # RedShift (need to fix insert for non-AWS) details <- createConnectionDetails(dbms = 'redshift',
-  # user = Sys.getenv('CDM5_REDSHIFT_USER'), password =
-  # URLdecode(Sys.getenv('CDM5_REDSHIFT_PASSWORD')), server = Sys.getenv('CDM5_REDSHIFT_SERVER'),
-  # schema = Sys.getenv('CDM5_REDSHIFT_CDM_SCHEMA')) connection <- connect(details)
-  # insertTable(connection = connection, tableName = 'person', data = data.frame(person_id = 1),
-  # dropTableIfExists = TRUE, createTable = TRUE, tempTable = FALSE) tables <-
-  # getTableNames(connection, Sys.getenv('CDM5_REDSHIFT_CDM_SCHEMA')) expect_true('PERSON' %in% tables)
-  # disconnect(connection)
+  # RedShift --------------------------------------------------
+  details <- createConnectionDetails(dbms = "redshift", 
+                                     user = Sys.getenv("CDM5_REDSHIFT_USER"), 
+                                     password = URLdecode(Sys.getenv("CDM5_REDSHIFT_PASSWORD")),
+                                     server = Sys.getenv("CDM5_REDSHIFT_SERVER"))
+  connection <- connect(details)
+  tables <- getTableNames(connection, Sys.getenv("CDM5_REDSHIFT_CDM_SCHEMA"))
+  expect_true("PERSON" %in% tables)
+  disconnect(connection)
 })